requirements.txt
alabaster==0.7.12
asgiref==3.8.1
Babel==2.10.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.4
Django==5.1
djangorestframework==3.15.2
docutils==0.17.1
gunicorn==20.1.0
idna==3.3
imagesize==1.3.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pygments==2.12.0
pyparsing==2.4.7
pytz==2021.3
requests==2.28.1
sentry-sdk==1.8.0
setuptools==73.0.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2024.1
urllib3==1.26.9
